requirements.txt
requests>=2.31.0
pandas>=2.0.0
PyYAML>=6.0.1
numpy>=1.24.0
hmmlearn>=0.3.0
scikit-learn>=1.3.0
python-dotenv
